Fixed: Remaining size_t printouts

+++ contrib/extractor/System.cpp
@@ -227,7 +227,7 @@ uint32 ReadMapDBC()
         map_ids[x].id = dbc.getRecord(x).getUInt(0);
         strcpy(map_ids[x].name, dbc.getRecord(x).getString(1));
     }
-    printf("Done! (%zu maps loaded)\n", map_count);
+    printf("Done! (%u maps loaded)\n", (unsigned int)map_count);
     return map_count;
 }
 
@@ -252,7 +252,7 @@ void ReadAreaTableDBC()
 
     maxAreaId = dbc.getMaxId();
 
-    printf("Done! (%zu areas loaded)\n", area_count);
+    printf("Done! (%u areas loaded)\n", (unsigned int)area_count);
 }
 
 void ReadLiquidTypeTableDBC()
@@ -273,7 +273,7 @@ void ReadLiquidTypeTableDBC()
     for (uint32 x = 0; x < LiqType_count; ++x)
         LiqType[dbc.getRecord(x).getUInt(0)] = dbc.getRecord(x).getUInt(3);
 
-    printf("Done! (%zu LiqTypes loaded)\n", LiqType_count);
+    printf("Done! (%d LiqTypes loaded)\n", (unsigned int)LiqType_count);
 }
 
 //
